+++ src/renderer/app/components/Activitybar/index.tsx
@@ -4,6 +4,7 @@ import { observer } from 'mobx-react-lite';
 import store from '~/renderer/app/store';
 import { SidebarContent } from '~/renderer/app/store/activitybar';
 import { StyledActivitybar, StyledItem } from './style';
+import { icons } from '~/renderer/constants/icons';
 
 interface ItemProps {
   content: SidebarContent;
