+++ src/renderer/app/components/Sidebar/Sites/index.tsx
@@ -3,11 +3,11 @@ import { observer } from 'mobx-react-lite';
 
 import store from '~/renderer/app/store';
 import { ISite } from '~/interfaces';
-import { icons } from '~/renderer/constants';
-import { formatSiteDescription } from '~/renderer/app/utils';
 import { ToolbarButton } from '~/renderer/components/ToolbarButton';
 import { StyledPage, Header } from '../style';
 import { StyledSite, Label, Description } from './style';
+import { formatSiteDescription } from '~/renderer/app/utils/site';
+import { icons } from '~/renderer/constants/icons';
 
 const onAdd = () => {
   store.dialog.show('add-site');
